Updated Action Bar
Updated app theme and AddEditActivity to make a properly functioning ActionBar, with a fully functioning Home button.

Also added a Cancel and a Delete menu buttons to the layout, and implemented the Home and Cancel logic.
Still need to implement Delete logic, after implementing the activity opening by clicking a food item in the first place.

The next steps begin with implemeting the following behaviour:
If the screen is small / horizontal, the layout reaches over the screen's limits.
In this case, the user should be able to scroll down to see the rest of the layout's content (category/date/time/etc).
When scrolling down, the ActionBar fades out, as seen in the Play store.