Opened AddEdit from food item
Enabled opening an AddEditActivity by clicking a food item in the InnerRecycler.

Accomplished that by using this guide:
http://bit.ly/2ycG46w
Essentialy by turning the FoodItemViewHolder from a dumb VH to a smart VH - simplifying its content binding, creating an inner interface which was implemented in MainActivity to handle food item clicks/long-clicks, and binding OnClickListeners to the view.
Through said interface the clicked food item was passed to the AddEditActivity to be edited.

As of now the contents of the food item are not yet displayed in the add_edit layout.

Current objectives:
* When activity is opened by an item, display its contents for editing
* Add DB functionality to the layout - adding and editing.
* Review the comments in the link provided above, to find possible improvements on this commit.
* Also for learning more about StringDef and DiffUtil!!!!
* In AddEditRecyclerViewAdapter, replace the RecyclerItemClickListener implementation with the logic explained above + when editing an item, change the send icon to a save icon.

P.S:
A simpler version of a ViewHolder listener serving to open an AddEditActivity is still in the code inside InnerRecyclerAdapter.java, commented out, for future reference.