Home button functionality
In the process of making the home button in AddEdit return to the previously opened tab instead of the first one.
Next TODO to try is in utilStartAddEditActivity method in MainActivity.
Also removed (commented out) the cancel button - consideration is required on whether and how to use it.

Next objectives:
* Finish home button functionality.
* Replace home button icon with X (if achieved, eliminates the need for a cancel button).
* Get rid of add_edit_layout file without breaking FAB.
* Revise all SwapCursor methods to more accurately handle database changes (replace notifyDatasetChanged with action-specific notification methods).