Replace AddEditFragment
This commit contains the code attempting to replace the way I pass data from AddEditActivity to AddEditFragment - instead of using the PassActivityDataToFragment interface, I replace the default-created fragment with one that's got the relevant data in a Bundle.
Sadly, that doesn't work as expected (old fragment's FAB is still visible), so after this commint I retire the fragment and move all its content to the AddEditActivity.

Also made finished several small TODOs.