Categories table is prepared (Udemy version)
Updated AppDatabase and AppProvider to handle a Categories table in the DB, in the fashion of Tim Buchalka's course.

I still need to make the Categories table pre-defined for the app.
I need to either research on how to accomplish this, or try to use the SQLiteAssetHelper library in Github ( https://github.com/jgilfelt/android-sqlite-asset-helper ) to do the same.