Resolved #2 - only month-specific items displayed
Pre-commit, every cardview would display the entire month's (tab's) food items.
Now every card only contains the food items relevant to the date represented by it.
This was accomplished by letting the InnerAdapter know which day of the month it represents. Then, on binding content to the viewholders, a check is performed - if the content to be filled is not from the same day, the viewholder is destroyed.
Also fixed some magin issues in row.xml files.