Load tab only when selected
Added implementation for only loading a month's tab when it's selected, instead of loading all of them (by default it was loading the first 2) on app start-up.
This was accomplished by adding 2 boolean variables to PageFragment, mIsStarted & mIsVisible, and loading the data according to their, while relying on onStart, onStop and setUserVisibleHint.

Also added some minor improvements:
* Replaced all Toasts with Snackbars for future usability.
* Fixed a bug where the app would randomly throw an NPE in FoodItemViewHolder.destroy.
* Added space between cardviews.