Tabs amount+content according to DB no.4
Removed DB active synchronization. Now app only loads the tabs from the DB once at startup.
This requires some work, as adding items that do not belong in an existing tab does not refresh the tabs (re-opening the app is required to refresh and see the new item).
However, this work is postponed until after the current MainActivity is changed to HistoryActivity.

Also updated some of the visuals (category colors, text formatting, style and margins, etc).