Code and tabs organization
Moved activity_main's content to content_main (seems reasonable, file names considered).
Fixed logic in MonthsStatePagerAdapter so that the tabs labeled "THIS MONTH" and "LAST MONTH" are not the last ones, but the ones matching this and the last months.
Set the default tab to be this month instead of the last one.