MainActivity changes
Changed MainActivity so it only displays the current month, without tabs.
The previous MainActivity file was renamed and revised as HistoryFragment (revisions still required). It's new purpose is to display all food items previous to this month, in tabs (still need to consider whether the current month should be included).
Fiddled around a little in the Manifest, because of errors brought on by ill-matching activity themes.

Next objectives:
* Finish all revisions on HistoryFragment,
* Add NavigationDrawer.