Continued migrating to Room no. 2
Created FoodDaoTest instrumentation test class. Began testing Dao, need to continue.
Added LiveDataTestUtil class from AOSP to assist in testing the Dao, because LiveData.getValue is asynchronous, meaning the test methods cannot properly receive a value from the db (NPE occur because the getValue method hasn't finished by the time asserts are called).
This is where LiveDataTestUtil comes into play. It only has one method - getValue(LiveData) - which synchronously queries the value contained by the LiveData parameter.

Formatted documentation and class names.
Due to FoodItem's name change to SectionChildFood, and DateHeader's name change to SectionHeaderDate, a lot of the code in the activities and fragments will have to be edited to be logical.
Most of the code will be revised anyway when implementing MVVM so that's not a big concern.

Next objectives:
* Finish migrating to Room (continue instrumentation tests then Cursor-related code cleanup and refactoring), then implement MVVM pattern.
* Improve images display for navigation drawer header (reduce in size OR fetch via API to reduce apk size).
* Implement Categories with icons.
* Create missing fragments from navigation drawer menu.